requirements.txt
streamlit
langchain 
langchain-community 
langchain-groq
pandas
python-dotenv